Upload getTriangle.ps1 file.
This is the first iteration of a .ps1 function that is meant to give you the sides of a triangle when you provide the fractional value of a trigonometric function. If only the numerator is given, the assumed denominator is 1. There are serious bugs and incomplete code, but I wanted to get something posted, to overcome my fear of publishing unfinished work (the work is never finished!).